+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `xan plot -A/--aggregate` does not take an expression anymore but has an automatic selection of two modes: `sum` and `mean`. It should also be faster.
 * Renaming the `index` function as `row_index` for clarity.
 * `xan agg -C/--along-columns & -M/--along-matrix` & `xan groupby -C/--along-columns & -M/--along-matrix` will not map current column index to the result of the `index()` function. The `col_index()` can be now used instead for this very purpose.
+* `xan window -g/--groupby` does not require the file to be sorted anymore. This means using `-g/--groupby` will now require the whole file to be buffered into memory by the command. The old behavior can still be used through the `-S/--sorted` flag, thus aligning the `xan window` command with the rest of the tool.
 
 *Features*